// src/app/models/programacion-model.ts
export class ProgramacionModel {
    idsEmpleados: number[];
    idArea: number
    fechaFin: string
    fechaInicio: string
    horaInicio: string
    horaFin: string
    nombreTurno: string

    idEmpleado: number
    fechaHoraInicio: string
    fechaHoraInicioBreak: string
    fechaHoraFinBreak: string
    fechaHoraFin: string
    nombreEmpleado: any;
    tipo: number;
    idProgramacion: number;
    nombreArea: String;

    turno:any

}